2024/otel-proto/pyproject.toml
[project]
name = "otel-proto"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Inada Naoki", email = "songofacandy@gmail.com" }
]
dependencies = [
    "opentelemetry-proto>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "otel_proto:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/otel_proto"]
